backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database (SQLite works out of the box, no extra driver needed)
sqlalchemy==2.0.25
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL
alembic==1.13.1

# AI/ML — Google Gemini (required for real analysis)
google-generativeai>=0.3.2

# Optional heavy ML dependencies (uncomment if you want CNN deepfake detection)
# torch>=2.2.0
# torchvision>=0.17.0
# sentence-transformers==2.3.1
# transformers==4.37.1
# pinecone-client==3.0.2
# openai-whisper>=20231117

# Image/Video processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Audio processing (optional)
# librosa==0.10.1
# soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
numpy>=1.24.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
